Move DerivationResult to a separate file outside mix-in
